refactor: minor changes
